+++ src/components/App/index.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import Header from 'components/Header';
-import Table from 'containers/Table';
-import Calendar from 'containers/Calendar';
-import List from 'containers/List';
+import Table from 'components/Table';
+import Calendar from 'components/Calendar';
+import List from 'components/List';
 import Footer from 'components/Footer';
 
 import './index.scss';
